backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
torch
PyPDF2
python-multipart
numpy
